requirement.txt
Flask
Pillow
tensorflow
tensorflow_hub
